+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2.3
 * Added support to fetch toggles from cache if got connection issue
+* Added support to update unleash context
 
 # 0.2.2
 * Added support to override fallback value